chore: ruff + mypy + prettier configs and apply lint fixes
Phase 1 of docs/IMPROVEMENT-PLAN.md — wires up the quality gates that CI
will enforce, and clears the existing code under them.

Backend (pyproject.toml)
  - Adds dev extras: ruff 0.15, mypy 2.1, pytest-cov 6, types-passlib
  - Ruff config: select E,F,W,I,B,UP,SIM,ASYNC,RUF; ignore E501, B008
    (FastAPI Depends pattern), SIM117, RUF002 (em-dashes), UP042
    (str-enum migration deferred — risks behaviour change before tests
    are tightened)
  - Mypy: strict_optional, warn_return_any, warn_unused_ignores,
    pydantic plugin. Not full --strict yet — Phase 6 tightens.
  - Coverage config: branch coverage, omits seed_db.py

Fixes (4 real, 24 auto-fixed by `ruff --fix`)
  - alembic/env.py: simplify _include_name to a single return (SIM103)
  - app/main.py: pass CORS kwargs directly instead of dict unpacking
    that mypy couldn't reconcile with the typed signature
  - app/core/security.py: explicit `dict[str, Any]` annotation on the
    jwt.decode return (no-any-return)
  - app/services/carbon.py: explicit `float` annotations through the
    elasticity calc; cast final product (no-any-return)
  - app/services/estimator.py: drop two unused `type: ignore` comments
  - app/services/providers.py: drop unnecessary int() around round()
  - app/schemas/ingestion.py: rejected default → Field(default_factory)
  - tests/test_api.py: lambda → def for the `transport` helper
  - 24 auto-fixes (UP017 datetime.UTC alias, I001 import order,
    RUF100 unused noqa, UP037 quoted-annotation, RUF022 dunder_all)

Frontend
  - .prettierrc.json + .prettierignore (with prettier-plugin-tailwindcss)
  - npm scripts: `format`, `format:check`
  - Reformat all of src/ (most diffs are whitespace + Tailwind class
    sort from the plugin)
  - Refactor one ternary in OnboardingFlow.tsx where Prettier and a
    line-comment fought; equivalent JSX render (null → <></>)

Verification (local, before committing the workflow)
  - backend: ruff check . → clean; mypy app → clean;
    pytest -q → 25 passed; coverage 71%
  - frontend: prettier --check → clean; next lint → clean;
    npm run typecheck → clean; npm run build → clean

+++ backend/app/services/providers.py
@@ -100,7 +100,7 @@ class SandboxBankProvider:
                         ProviderTxn(
                             external_id=f"sbx-{seq}",
                             booked_at=booked,
-                            amount_minor=int(round(pounds * 100)),
+                            amount_minor=round(pounds * 100),
                             currency="GBP",
                             mcc=mcc,
                             description=f"{merchant} card purchase",
